# app/Http/Controllers/api/ForgotPasswordController.php
<?php

namespace App\Http\Controllers\api;

use App\Http\Controllers\Controller;
use Illuminate\Http\Request;
use Illuminate\Support\Facades\Mail;
use App\Mail\SendCodeResetPassword;
use Illuminate\Mail\Mailable;
use App\Models\ResetCodePassword;


class ForgotPasswordController extends Controller
{

     /**
     * @OA\Post(
     * path="/api/password/email",
     * summary="forgot password",
     * description="Enter email to reset password !",
     * operationId="ForgotPassword",
     * security={{"bearer_token":{}}},
     * tags={"Forgot password"}, 

     *     @OA\Parameter(
     *          name="email",
     *          description="Enter name",
     *          required=true,
     *          in="query",
     *          @OA\Schema(
     *              type="string"
     *          )
     *       ),
     *
     *      @OA\Response(
     *          response=200,
     *          description="Valid email ."
     *     ),
     *      @OA\Response(
     *          response=204,
     *          description="Successful operation",
     *          @OA\JsonContent()
     *       ),
     *      @OA\Response(
     *          response=400,
     *          description="Wrong Input email",
     *      ),
     *      @OA\Response(
     *          response=401,
     *          description="Unauthenticated",
     *      ),
     *      @OA\Response(
     *          response=403,
     *          description="Forbidden"
     *      ),
     *      @OA\Response(
     *          response=404,
     *          description="Resource Not Found"
     *      )
     * )
     */

    public function __invoke(Request $request)
    {
        $data = $request->validate([
            'email' => 'required|email|exists:users',
        ]);

        // Delete all old code that user send before.
        ResetCodePassword::where('email', $request->email)->delete();

        // Generate random code
        $data['code'] = mt_rand(100000, 999999);

        // Create a new code
        $codeData = ResetCodePassword::create($data);

        // Send email to user
        Mail::to($request->email)->send(new SendCodeResetPassword($codeData->code));

        return response()->json([
            'message ' => trans('passwords.sent')], 200);

    }
}